LayoutTests/imported/w3c/web-platform-tests/css/css-text/animations/word-boundary-detection-interpolation.html

<!DOCTYPE html>
<meta charset="UTF-8">
<link rel="help" href="https://drafts.csswg.org/css-text-4/#word-boundaries">
<meta name="test" content="word-boundary-detection supports animation">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/interpolation-testcommon.js"></script>
<style>
.parent {
  text-indent: 70px;
}
.target {
  font: 10px sans-serif;
  text-indent: 10px;
  margin-left: 20px;
}
.expected {
  color: green;
}
</style>

<body>
  <template id="target-template">T</template>
</body>

<script>
test_no_interpolation({
  property: 'word-boundary-detection',
  from: 'auto("en_US")',
  to: 'normal',
});
</script>
